feat: add initial multipage support
